removed register activity.
