+++ RatatouilleChefApp/app/src/main/res/navigation/nav_graph.xml
@@ -28,5 +28,9 @@
         android:id="@+id/registerFragment"
         android:name="edu.javeriana.ratatouille_chef_app.authentication.ui.RegisterFragment"
         android:label="fragment_register"
-        tools:layout="@layout/fragment_register" />
+        tools:layout="@layout/fragment_register">
+        <action
+            android:id="@+id/action_registerFragment_to_loginFragment"
+            app:destination="@id/loginFragment" />
+    </fragment>
 </navigation>
